lcd model generator: modified model to cope with list parameters
Since INI backend is more strict, the old example broke. This is due to
LCDproc Hello and GoodBye parameters which are repeated in the example file.
Before only the first value was stored in the dummy model. This is
no longer possible and the dummy containing the raw LCDproc.conf
infor model must now provide list array for Hello and GoodBye.