Fix gen_class_pod which skipped some classes
In more details, when generating doc for OpenSsh model, gen_class_pod
was skipping Ssh class , because this class was inluded by SystemSsh and
the test used to avoid generating twice the same class was wrong.

An included class is known by config::model by no doc is generated for
an included class.

This should fix the reproducible build issue seen in Debian:
https://tests.reproducible-builds.org/dbd/unstable/i386/libconfig-model-openssh-perl_1.237-1.diffoscope.html