move warp info in warp param in WarpedNode...
In WarpedNode, warp information used to be specified in
rules and follow parameter.

On one hand this made sense as rules is mandatory for WarpedNode.
(A WarpedNode without rules is just a plain Node).

On the other hand, warp specification is not consistent with
similar spec for Value, Id or Hash element and makes Itself model
more complex.

So directly specifying rules and follow is deprecated in favor of
using warp => { follow => {...}, rules => [...] ) just like the
other elements