fix shell set command to accept white spaces
previous commit did change the parameters passed to set subref.
All args are now passed as an array instead of a string.

Joining the args is a way to remove the white spaces around = and :

But this works only for a single set command. ie.

set foo=bar foo2=bar2

is broken. OTOH, I doubt that it's used.

This is still a kludge. May be I should avoid using load and simply
accept

set key [=] value

this would also enable cd in a hash or an array.

A separate load command would still be possible...