plainfile backend: improve file specification.
Allow more latitude to name the file using a formula. Now the functions
are computed from the saved object and not from the parent node. This
enables using the element name in the file name.

On the other hand the computation of the formula is slightly different
compared to the computation done by backend mgr which is done starting
from the node attached to the backend manager. This cannot be avoided
since the backend manager is not attached to a simple object.