clean up check on node init (gh #15)
During init, data are read from config files  for the first time. At
this stage, we must use the check setting passed when creating the
instance. This way, the check parameter used when creating the instance
is always used to read data from file.

This was the cause of the dropped data seen in gh #15: the file below
was read with check => yes instead of check => no