use logger to warn issues...
when $::_use_log4perl_to_warn is set. This global variable will be set
in cme and in all tests that check emitted warnings. Using this variable
avoids breaking all tests in other Config::Model modules. It will be
removed once the tests of all other modules are ported to the new test
system.