allow pattern to define help on value
I.e. the help on value is specified with a pattern attached to the
beginning of the value. The help is shown in the GUI when the current
value matches the pattern.

The hash key that define the pattern are sorted according to their
length so the most specific patterns are tested first. This also allow
model designer to specify an empty pattern that is used as a fall-back.

I.e
 help => {
   'foothis' => 'help for foothis, foothis2...',
   'foo' => 'help for foothat but not for foothis',
   '' => help for anything else'
 }