Value: do not store warnings from computed value
Because the computed value may have a bad value which is overridden by
user. In this case any warning generated from the bad computed value
is misleading.

Furthermore, the old behavior was to clear the list of warning when
computing a value (where the result is good) even though the
overridden value is bad (kinda opposite effect of the previous
paragraph). This patch also fixes this problem.