fix (Node): undef warning on $filter value
Setting a default value with Perl signatures is not enough as cme
calls apply_fixes with (undef) most of the time. Perl checks only the
lenght of the @ARGV array before deciding to apply the default
value. So $filter was set to undef.

This fix no longet try to be smart with the regexp (by using /./ as an
always true regexp)