fix (Value): consider and empty string as an unset value
This is required to fix a bug with Dpkg model.

Vcs-Browser element is a computed value (a default value that depends
on other elements) that can be overridden.

If this overridden value is an empty string, the value shows up as
empty even though it should be filled with a default value.

Furthermore, the distinction between undef and empty value is too
subtle for a configuration file so an empty string is considered as
unset and replace with its default or computed value (if any)