fix (Value): do not check compute result with mandatory value
This does not make sense because a computed value can be a default
value. If the computation returns undef, it means that there's no
default value and we must give the user the oppotunity to set this
value.

This was not the case before where mandatory was also applied to
computed result.

See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1015913