HashId: improve copy method to fix PlainFile backend
The problem was that plainfile backend was not able to remove files
when needed.

This problem was seen when renaming of install file =libuv1.install=
to =libuv1t64.install= via =cme edit dpkg=. The operation did not lead
to the rename of the install files.

See the new test to check this condition. (This test emulates the dpkg
model used to handle *.install files).

The trick is to:
- copy data: so that a new PLainFile backend is created to write the
new file
- remove the old node. Since its write backend was registered in the
instance, the obsolete write backend sub was called. This one's detect
that it's supporting node is gone and remove the associated file.